Hourly forecast (#17)

+++ src/main/java/com/jonjam/weathermcp/currentconditions/CurrentConditionsDto.java
@@ -14,9 +14,9 @@ public class CurrentConditionsDto {
 
   String weatherText;
 
-  long temperatureMetric;
+  float temperatureMetric;
 
-  long temperatureImperial;
+  float temperatureImperial;
 
   String link;
 }
